requirements.txt
django==5.0.2
djangorestframework==3.14.0
djongo==1.3.6
pymongo==3.12.3
channels==4.0.0
python-dotenv==1.0.1
requests==2.31.0
django-cors-headers==4.3.1 